First commit of the project with initial tested version
This is a simple project to properly notify in Slack any budget alerts
for GCP. Since using PubSub as messaging queue for these alerts are prone
to multiple messages a day and even disregarding order.

The main idea is to use Firebase in its free tier as a simples database
to store last notified threshold and only send Slack message when a new
threshold is reached.

Currently supported thresholds are 50%, 75% and 100% only.

Signed-off-by: Ricardo Verhaeg <ricardo@arquivei.com.br>